requirements.txt
streamlit
mysql-connector-python
pandas
prophet
groq
python-dotenv
plotly